backend/requirements.txt
asgiref==3.6.0
astroid==3.0.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
chardet==5.2.0
contourpy==1.1.0
crispy-bootstrap5==0.7
cycler==0.11.0
dill==0.3.7
Django==4.2.1
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-countries==7.5.1
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-filter==23.2
django-seed==0.3.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
et-xmlfile==1.1.0
Faker==18.11.1
fonttools==4.41.1
importlib_resources==6.4.5
isort==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.2
mccabe==0.7.0
numpy==1.24.4
openpyxl==3.1.2
packaging==23.1
Pillow==9.5.0
platformdirs==3.11.0
psycopg2-binary>=2.9
pylint==3.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2024.1
reportlab==4.2.5
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
toposort==1.10
typing_extensions==4.5.0
zipp==3.20.2
